Change max_tokens from u16 to u32 for larger context window in recent models #172

+++ src/language_models/options.rs
@@ -7,7 +7,7 @@ use crate::schemas::{FunctionCallBehavior, FunctionDefinition};
 #[derive(Clone)]
 pub struct CallOptions {
     pub candidate_count: Option<usize>,
-    pub max_tokens: Option<u16>,
+    pub max_tokens: Option<u32>,
     pub temperature: Option<f32>,
     pub stop_words: Option<Vec<String>>,
     pub streaming_func: Option<
@@ -56,7 +56,7 @@ impl CallOptions {
     }
 
     // Refactored "with" functions as methods of CallOptions
-    pub fn with_max_tokens(mut self, max_tokens: u16) -> Self {
+    pub fn with_max_tokens(mut self, max_tokens: u32) -> Self {
         self.max_tokens = Some(max_tokens);
         self
     }
